todo/requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.63.0
greenlet @ file:///tmp/build/80754af9/greenlet_1611958330323/work
h11==0.12.0
httptools==0.1.1
passlib==1.7.4
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
python-dotenv==0.17.0
python-jose==3.2.0
PyYAML==5.4.1
rsa==4.7.2
six==1.15.0
SQLAlchemy @ file:///tmp/build/80754af9/sqlalchemy_1618089536876/work
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
